+++ src/http/message/live/play.rs
@@ -11,6 +11,6 @@ pub struct LivePlayResponse {
     pub code: u32,
     pub msg: String,
     pub gb_code: String,
-    pub stream_id: u64,
+    pub stream_id: u32,
 }
