+++ src/http/message/replay/keep_alive.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize)]
 pub struct ReplayKeepAliveRequest {
     pub gb_code: String,
-    pub stream_id: u64,
+    pub stream_id: u32,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -12,6 +12,6 @@ pub struct ReplayKeepAliveResponse {
     pub code: u32,
     pub msg: String,
     pub gb_code: String,
-    pub stream_id: u64,
+    pub stream_id: u32,
 }
