+++ src/http/message/replay/start.rs
@@ -13,6 +13,6 @@ pub struct ReplayStartResponse {
     pub code: u32,
     pub msg: String,
     pub gb_code: String,
-    pub stream_id: u64,
+    pub stream_id: u32,
 }
